requirements.txt
pynacl>=1.5.0
cryptography>=41.0.0
stem>=1.8.2
PySocks>=1.7.1
prompt_toolkit>=3.0.43
rich>=13.7.0
